server/requirements.txt
flask==1.1.1
graphene==2.1.8
flask_graphql==2.0.1
flask-cors==3.0.8
